+++ sphinx-docs/source/changelog.rst
@@ -4,9 +4,14 @@ Change Log
 v2.4 (TBD)
 --------------------
 
+*New Features*
+
+- Can compute mapped nlists/positions while running a simulation
+
 *Enhancements*
 
 - Added benchmarking to CI and reporting
+- Various documentation improvements
 
 v2.3 (2021-04-06)
 --------------------